Move tests for exponential backoff
